+++ jest.config.js
@@ -23,6 +23,6 @@
 module.exports = {
   preset: 'ts-jest',
   testEnvironment: 'node',
-  testPathIgnorePatterns: ['/node_modules/', '/__tests__\\/_.*'],
+  testPathIgnorePatterns: ['/node_modules/', '/__tests__\\/__utils__/'],
   setupFiles: ['dotenv/config'],
 }
