+++ src/graphql/data-sources/index.ts
@@ -19,7 +19,7 @@
  * @license   http://www.apache.org/licenses/LICENSE-2.0 Apache License 2.0
  * @link      https://github.com/serlo-org/api.serlo.org for the canonical source repository
  */
-import { GoogleSheetApi } from './google-spreadsheet'
+import { GoogleSheetApi } from './google-spreadsheet-api'
 import { SerloDataSource } from './serlo'
 
 export interface DataSources {
